feat(analytics): bound click tracking by distinct links
- Aggregate clicks for up to 10,000 distinct links.
- Preserve incoming clicks during flushes and db failures.
- Serialize flushes to prevent duplicate writes during shutdown.
- Migrate click counts to bigint and update API to use long.
- Add concurrency, recovery, capacity, and large-count tests.

+++ UrlShortener.Api/Models/Link.cs
@@ -18,7 +18,7 @@ public class Link
     [MaxLength(7)]
     public required string ShortCode { get; set; }
 
-    public int ClickCount { get; set; }
+    public long ClickCount { get; set; }
 
     [DataType(DataType.Date)]
     public DateTime CreatedAt { get; init; } = DateTime.UtcNow;
